Code adapted to upgraded encoding system
(1) Rhythms involving sixteenth notes are now read correctly.
(2) The ybcNote class has a new data attribute .restDot (boolean) that
indicates whether a note with determinate pitch was notated as a dot
(i.e. a rest understood to have the pitch value of the previous note).
(3) Instances of ybcNote and ybcGatra can now be printed